+++ packages/pact-http-controller/c8yctrl-dev.config.ts
@@ -6,7 +6,7 @@ import {
   C8yPactHttpController,
   C8yPactHttpControllerConfig,
   C8yPactHttpResponse,
-} from "cumulocity-cypress/node";
+} from "../../dist/bin/c8yctrl";
 
 import { Request } from "express";
