+++ packages/pact-http-controller/contrib/c8yctrl.config.ts
@@ -2,7 +2,7 @@ import _ from "lodash";
 import fs from "fs";
 import path from "path";
 
-import { C8yPactHttpControllerConfig } from "cumulocity-cypress/node";
+import { C8yPactHttpControllerConfig } from "cumulocity-cypress/shared/c8yctrl";
 
 import { createLogger, format, transports } from "winston";
 // https://github.com/winstonjs/winston/issues/2430
