+++ src/index.ts
@@ -1,11 +1,7 @@
-export * from "./shared/c8ypact";
-export * from "./lib/pact/runner";
-export * from "./lib/pact/fetchclient";
-export * from "./lib/pact/cypresspact";
 export * from "./shared/auth";
-export * from "./shared/c8ypact/fileadapter";
 export * from "./shared/versioning";
-export { registerDefaultLocales, registerLocale } from "./lib/locale/locale";
+export * from "./shared/c8ypact/schema";
+
 export {
   isCypressError,
   isIResult,
