--- src/lib/commands/c8ypact.ts
@@ -1,4 +1,0 @@
-import "./requires";
-import "../pact/cypresspact";
-import "../pact/c8ymatch";
-import "../pact/fetchclient";
