--- src/node.ts
@@ -1,13 +1,0 @@
-export * from "./shared/c8ypact";
-export * from "./shared/auth";
-export * from "./shared/c8ypact/fileadapter";
-export * from "./shared/c8yctrl";
-export {
-  isCypressError,
-  isIResult,
-  isWindowFetchResponse,
-  toCypressResponse,
-  C8yClient,
-  C8yClientOptions,
-  oauthLogin,
-} from "./shared/c8yclient";
