+++ src/shared/c8ypact/index.ts
@@ -4,3 +4,14 @@ export * from "./c8ydefaultpactrecord";
 export * from "./schema";
 export * from "./matcher";
 export * from "./preprocessor";
+export * from "./fileadapter";
+
+export {
+  isCypressError,
+  isIResult,
+  isWindowFetchResponse,
+  toCypressResponse,
+  C8yClient,
+  C8yClientOptions,
+  oauthLogin,
+} from "./../c8yclient";
